+++ src/engine/mod.rs
@@ -3,14 +3,14 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 pub use self::{
-    engine::{BlockDev, Engine, Filesystem, Pool, Report, ReportType},
+    engine::{BlockDev, Engine, Filesystem, Pool, Report},
     event::{get_engine_listener_list_mut, EngineEvent, EngineListener},
     sim_engine::SimEngine,
     strat_engine::StratEngine,
     types::{
         BlockDevState, BlockDevTier, CreateAction, DeleteAction, DevUuid, EngineAction,
-        FilesystemUuid, MaybeDbusPath, Name, PoolUuid, Redundancy, RenameAction, SetCreateAction,
-        SetDeleteAction,
+        FilesystemUuid, MaybeDbusPath, Name, PoolUuid, Redundancy, RenameAction, ReportType,
+        SetCreateAction, SetDeleteAction,
     },
 };
